refactor(transfers): Extract Stepper styles
