+++ src/components/Stepper/styles.styl
@@ -1,0 +1,6 @@
+.ProgressBackground
+    background-color var(--silver)
+
+.ProgressLine
+    height 5px
+    margin-bottom 0.5rem
